script for selecting and filtering intensity over time
This script reads experimental data for the beam intensity and selects either high or low intensity beams while ignoring alpha rays. Via given start and end timestamp in UNIX time the intensity is selected from the PS public ROOT file. Additionally, the frequency of such intensities is being plotted for high and low intensity beams as well as in comparison to visualize the beams overall intensity for the given time to evaluate the PS proton beam supply