added script for time offset of PS and experiment due to missing internet connection and internal clock speed
The ROOT file provided by PS does not match the time recorded by the experiment. The difference of time between those gets progressively larger found by matching the bumps and hits through best matchmaking. After that the time difference per file is being evaluated and proven that the time gap between those two get larger. It is found to be at the beginning of the experiment to be around 600ms and by the end of the experiment (duration ~15 hours) to be a little over 900ms. Thiis was found to be due to the internal clock speed of the experiments recording computer which ist 6ns per s slower than UNIX official time (used by PS). The computer was last connected to network about 30 hours before the experiment start which would also match the shift of 600ms at the beginning of the experiment.

With the result of this linear progression the data is then newly matched and found to be fitting.