pulse intensity filter script added
Added pulse intensity filter script and example outputs.
This script matches the times from PS and the times from the experiment which where shifted by a fix rate due to internal clock speed differences. It then adds the pulse intensity which is possible through the mating of time and then splits the matches into high intensity and low intensity bunches. Those are then being split up into if their time of flight was higher or lower than 10ms and then plotted in comparison to see the average TOF for high and low intensity bunches